refactor(admin): split AddHeroesPanel into focused sibling modules
AddHeroesPanel.tsx had grown to 1,028 lines mixing the panel component,
its leaf row components, a large StyleSheet, and mode/stage constants.
Extract the self-contained pieces into siblings:

  AddHeroesPanel.styles.ts      the shared StyleSheet
  AddHeroesPanelRows.tsx        CharacterRow / MemberRow (+ Checkbox,
                                StatusBadge, charSub)
  AddHeroesPanel.constants.ts   AddedHero/Mode/Flash types + STAGE_BADGE,
                                MODES, GROUP_ICON, PLACEHOLDER

The shared styles live in one module imported by both the panel and the
rows, so nothing is duplicated. AddHeroesPanel.tsx drops 1,028 -> 655
lines. Pure code-move; no behavior change. typecheck clean; format
clean; 363 tests pass; 0 lint errors.
