+++ src/components/admin/health/AddHeroesPanel.styles.ts
@@ -1,0 +1,177 @@
+import { StyleSheet } from 'react-native';
+import { COLORS } from '../../../constants/colors';
+
+export const styles = StyleSheet.create({
+  modeRow: { flexDirection: 'row', flexWrap: 'wrap', gap: 6, rowGap: 6, marginBottom: 10 },
+  modePill: {
+    paddingHorizontal: 12,
+    paddingVertical: 6,
+    borderRadius: 8,
+    backgroundColor: '#efe6d6',
+  },
+  modePillOn: { backgroundColor: COLORS.navy },
+  modePillText: { fontFamily: 'Nunito_700Bold', fontSize: 12, color: COLORS.navy },
+  modePillTextOn: { color: '#fff' },
+
+  searchBox: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: 8,
+    backgroundColor: '#f6f0e6',
+    borderRadius: 10,
+    paddingHorizontal: 13,
+    paddingVertical: 10,
+  },
+  searchInput: { flex: 1, fontFamily: 'Nunito_700Bold', fontSize: 14, color: COLORS.black },
+  popularHint: {
+    fontFamily: 'Nunito_400Regular',
+    fontSize: 12.5,
+    color: COLORS.grey,
+    lineHeight: 18,
+  },
+  loadMore: { alignItems: 'center', justifyContent: 'center', paddingVertical: 12, marginTop: 4 },
+  loadMoreText: {
+    fontFamily: 'Nunito_700Bold',
+    fontSize: 12.5,
+    color: COLORS.orange,
+    textDecorationLine: 'underline',
+  },
+  dim: { opacity: 0.4 },
+
+  toolbar: { flexDirection: 'row', alignItems: 'center', gap: 10, marginTop: 12, flexWrap: 'wrap' },
+  backBtn: { flexDirection: 'row', alignItems: 'center', gap: 3, flex: 1, minWidth: 0 },
+  backText: { fontFamily: 'Nunito_700Bold', fontSize: 13, color: COLORS.navy, flexShrink: 1 },
+  countText: { fontFamily: 'Nunito_700Bold', fontSize: 12, color: COLORS.grey },
+  miniToggle: {
+    paddingHorizontal: 10,
+    paddingVertical: 5,
+    borderRadius: 7,
+    backgroundColor: '#efe6d6',
+  },
+  miniToggleOn: { backgroundColor: COLORS.navy },
+  miniToggleText: { fontFamily: 'Nunito_700Bold', fontSize: 11, color: COLORS.navy },
+  miniToggleTextOn: { color: '#fff' },
+  linkBtn: { paddingVertical: 4 },
+  linkText: {
+    fontFamily: 'Nunito_700Bold',
+    fontSize: 12,
+    color: COLORS.orange,
+    textDecorationLine: 'underline',
+  },
+  addBtn: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: 5,
+    backgroundColor: COLORS.orange,
+    borderRadius: 9,
+    paddingHorizontal: 13,
+    paddingVertical: 8,
+  },
+  addBtnText: { fontFamily: 'Nunito_700Bold', fontSize: 12, color: '#fff' },
+  buildAllBtn: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: 5,
+    backgroundColor: COLORS.navy,
+    borderRadius: 9,
+    paddingHorizontal: 13,
+    paddingVertical: 8,
+  },
+  buildAllText: { fontFamily: 'Nunito_700Bold', fontSize: 12, color: '#fff' },
+
+  scroll: { maxHeight: 300, marginTop: 8 } as object,
+  charWrap: { borderBottomWidth: 1, borderBottomColor: 'rgba(41,60,67,0.06)' },
+  charWrapOpen: {
+    backgroundColor: '#faf5ec',
+    borderRadius: 10,
+    borderBottomColor: 'transparent',
+    marginVertical: 2,
+  },
+  row: { flexDirection: 'row', alignItems: 'center', gap: 10, paddingVertical: 7 },
+  rowBody: { flex: 1, minWidth: 0, flexDirection: 'row', alignItems: 'center', gap: 10 },
+  groupRow: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: 10,
+    paddingVertical: 7,
+    borderBottomWidth: 1,
+    borderBottomColor: 'rgba(41,60,67,0.06)',
+  },
+  groupIcon: {
+    width: 34,
+    height: 44,
+    borderRadius: 7,
+    backgroundColor: COLORS.orange + '14',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  meta: { flex: 1, minWidth: 0, gap: 2 },
+  name: { fontFamily: 'Nunito_700Bold', fontSize: 14, color: COLORS.black },
+  sub: { fontFamily: 'Nunito_400Regular', fontSize: 12, color: COLORS.grey },
+  memberRow: { flexDirection: 'row', alignItems: 'center', gap: 10, paddingVertical: 6 },
+  memberBody: { flex: 1, minWidth: 0, flexDirection: 'row', alignItems: 'center', gap: 10 },
+  memberName: { flex: 1, fontFamily: 'Nunito_400Regular', fontSize: 13, color: COLORS.black },
+  empty: { fontFamily: 'Nunito_400Regular', fontSize: 13, color: COLORS.grey, marginTop: 12 },
+
+  cb: {
+    width: 19,
+    height: 19,
+    borderRadius: 5,
+    borderWidth: 2,
+    borderColor: 'rgba(41,60,67,0.3)',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  cbOn: { backgroundColor: COLORS.orange, borderColor: COLORS.orange },
+  cbDisabled: { borderColor: 'rgba(41,60,67,0.12)', backgroundColor: 'rgba(41,60,67,0.05)' },
+
+  badge: { flexDirection: 'row', alignItems: 'center', gap: 3 },
+  badgeText: { fontFamily: 'Nunito_700Bold', fontSize: 11 },
+
+  session: { marginTop: 12, gap: 8 },
+  loopBar: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: 8,
+    backgroundColor: COLORS.orange + '12',
+    borderRadius: 10,
+    paddingHorizontal: 12,
+    paddingVertical: 9,
+  },
+  loopText: { flex: 1, fontFamily: 'Nunito_700Bold', fontSize: 12.5, color: COLORS.navy },
+  loopBtn: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: 5,
+    backgroundColor: COLORS.orange,
+    borderRadius: 8,
+    paddingHorizontal: 12,
+    paddingVertical: 7,
+  },
+  loopBtnText: { fontFamily: 'Nunito_700Bold', fontSize: 12, color: '#fff' },
+  addedList: { gap: 2 },
+  addedRow: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: 10,
+    paddingVertical: 5,
+    borderBottomWidth: 1,
+    borderBottomColor: 'rgba(41,60,67,0.06)',
+  },
+  addedName: {
+    flex: 1,
+    minWidth: 0,
+    fontFamily: 'Nunito_700Bold',
+    fontSize: 13,
+    color: COLORS.black,
+  },
+  stageBadge: { fontFamily: 'Nunito_700Bold', fontSize: 11, fontVariant: ['tabular-nums'] },
+  removeBtn: {
+    width: 30,
+    height: 30,
+    borderRadius: 8,
+    alignItems: 'center',
+    justifyContent: 'center',
+    backgroundColor: COLORS.red + '12',
+  },
+});
