+++ src/components/admin/health/AddHeroesPanel.tsx
@@ -3,15 +3,7 @@
 // roster filter, a duplicate guard, and a one-click "enrich now" to close the
 // loop. Added heroes enter as 'pending' and flow into step 1.
 import { useEffect, useMemo, useState } from 'react';
-import {
-  View,
-  Text,
-  TextInput,
-  Pressable,
-  ActivityIndicator,
-  ScrollView,
-  StyleSheet,
-} from 'react-native';
+import { View, Text, TextInput, Pressable, ActivityIndicator, ScrollView } from 'react-native';
 import { Ionicons } from '@expo/vector-icons';
 import { COLORS } from '../../../constants/colors';
 import { Panel } from './Panel';
@@ -33,56 +25,18 @@ import {
   type GroupResource,
 } from '../../../lib/db/cvIngest';
 import { getBuildHeroes, type BuildStage } from '../../../lib/db/build';
-import { CharacterPreview } from './CharacterPreview';
 
-// A character added during this session — enough to show it, build it, or undo it.
-// heroId is the minted internal id (PK, used to build/delete/poll); comicvineId is
-// the source id (used to dedupe against ComicVine search results).
-type AddedHero = { heroId: string; comicvineId: string; name: string; image: string | null };
-
-// Live build-stage badge for an added hero (updates as the pipeline runs).
-const STAGE_BADGE: Record<BuildStage, { label: string; color: string }> = {
-  comicvine: { label: 'ComicVine', color: COLORS.orange },
-  resolve: { label: 'resolving', color: COLORS.orange },
-  appearances: { label: 'appearances', color: COLORS.orange },
-  review: { label: 'needs review', color: COLORS.yellow },
-  unresolved: { label: 'no match', color: COLORS.grey },
-  failed: { label: 'failed', color: COLORS.red },
-  done: { label: 'built', color: COLORS.green },
-};
-
-type Mode = 'name' | 'popular' | 'team' | 'volume' | 'person' | 'movie' | 'publisher' | 'power';
-type Flash = (msg: string, tone?: 'info' | 'success' | 'error' | 'pending') => void;
-const MODES: { key: Mode; label: string }[] = [
-  { key: 'popular', label: '★ Popular gaps' },
-  { key: 'name', label: 'By name' },
-  { key: 'team', label: 'By team' },
-  { key: 'volume', label: 'By series' },
-  { key: 'person', label: 'By creator' },
-  { key: 'movie', label: 'By film' },
-  { key: 'publisher', label: 'By publisher' },
-  { key: 'power', label: 'By power' },
-];
-// Group-mode icon + search placeholder per resource.
-type IconName = 'people' | 'book' | 'brush' | 'film' | 'business' | 'flash';
-const GROUP_ICON: Record<string, IconName> = {
-  team: 'people',
-  volume: 'book',
-  person: 'brush',
-  movie: 'film',
-  publisher: 'business',
-  power: 'flash',
-};
-const PLACEHOLDER: Record<Mode, string> = {
-  name: 'Character name… (e.g. Darth Vader)',
-  popular: '',
-  team: 'Team name… (e.g. Jedi Order)',
-  volume: 'Comic series… (e.g. Star Wars)',
-  person: 'Creator name… (e.g. Jack Kirby)',
-  movie: 'Film title… (e.g. The Empire Strikes Back)',
-  publisher: 'Publisher… (e.g. Valiant)',
-  power: 'Power… (e.g. Telepathy)',
-};
+import {
+  STAGE_BADGE,
+  MODES,
+  GROUP_ICON,
+  PLACEHOLDER,
+  type AddedHero,
+  type Mode,
+  type Flash,
+} from './AddHeroesPanel.constants';
+import { CharacterRow, MemberRow } from './AddHeroesPanelRows';
+import { styles } from './AddHeroesPanel.styles';
 
 export function AddHeroesPanel({
   flash,
@@ -699,330 +653,3 @@ export function AddHeroesPanel({
     </Panel>
   );
 }
-
-// The result sub-line: appearances + publisher when we have a popularity count
-// (name + popular feeds both carry it now), otherwise publisher or the deck.
-function charSub(c: CvCharacter): string {
-  if (c.appearances != null) {
-    return `${c.appearances.toLocaleString()} appearances${c.publisher ? ` · ${c.publisher}` : ''}`;
-  }
-  return c.publisher ?? c.deck ?? '—';
-}
-
-// One character result: checkbox (multi-select) + tappable body (expand preview).
-// The checkbox and the body are separate hit targets so ticking doesn't expand.
-function CharacterRow({
-  c,
-  selected,
-  inCat,
-  dup,
-  expanded,
-  detail,
-  detailLoading,
-  busy,
-  onToggleSelect,
-  onToggleExpand,
-  onAdd,
-}: {
-  c: CvCharacter;
-  selected: boolean;
-  inCat: boolean;
-  dup: boolean;
-  expanded: boolean;
-  detail: CvCharacterDetail | null | undefined;
-  detailLoading: boolean;
-  busy: boolean;
-  onToggleSelect: () => void;
-  onToggleExpand: () => void;
-  onAdd: () => void;
-}) {
-  return (
-    <View style={[styles.charWrap, expanded && styles.charWrapOpen]}>
-      <View style={styles.row}>
-        <Pressable onPress={onToggleSelect} disabled={inCat} hitSlop={8}>
-          <Checkbox checked={selected} disabled={inCat} />
-        </Pressable>
-        <Pressable onPress={onToggleExpand} style={styles.rowBody}>
-          <HeroThumb uri={c.image} width={32} height={42} radius={6} />
-          <View style={styles.meta}>
-            <Text style={styles.name} numberOfLines={1}>
-              {c.name}
-            </Text>
-            <Text style={styles.sub} numberOfLines={1}>
-              {charSub(c)}
-            </Text>
-          </View>
-          <StatusBadge inCat={inCat} dup={dup} />
-          <Ionicons name={expanded ? 'chevron-up' : 'chevron-down'} size={16} color={COLORS.grey} />
-        </Pressable>
-      </View>
-      {expanded ? (
-        <CharacterPreview
-          detail={detail}
-          loading={detailLoading}
-          fallbackImage={c.image}
-          inCat={inCat}
-          busy={busy}
-          onAdd={onAdd}
-        />
-      ) : null}
-    </View>
-  );
-}
-
-// A roster member (team / series / creator / film / publisher / power). Members
-// arrive as just id + name, so the collapsed row is compact — tapping the body
-// expands the same preview (its detail is fetched lazily from the id).
-function MemberRow({
-  m,
-  selected,
-  inCat,
-  dup,
-  expanded,
-  detail,
-  detailLoading,
-  busy,
-  onToggleSelect,
-  onToggleExpand,
-  onAdd,
-}: {
-  m: { id: string; name: string };
-  selected: boolean;
-  inCat: boolean;
-  dup: boolean;
-  expanded: boolean;
-  detail: CvCharacterDetail | null | undefined;
-  detailLoading: boolean;
-  busy: boolean;
-  onToggleSelect: () => void;
-  onToggleExpand: () => void;
-  onAdd: () => void;
-}) {
-  return (
-    <View style={[styles.charWrap, expanded && styles.charWrapOpen]}>
-      <View style={styles.memberRow}>
-        <Pressable onPress={onToggleSelect} disabled={inCat} hitSlop={8}>
-          <Checkbox checked={selected} disabled={inCat} />
-        </Pressable>
-        <Pressable onPress={onToggleExpand} style={styles.memberBody}>
-          <Text style={styles.memberName} numberOfLines={1}>
-            {m.name}
-          </Text>
-          <StatusBadge inCat={inCat} dup={dup} />
-          <Ionicons name={expanded ? 'chevron-up' : 'chevron-down'} size={16} color={COLORS.grey} />
-        </Pressable>
-      </View>
-      {expanded ? (
-        <CharacterPreview
-          detail={detail}
-          loading={detailLoading}
-          fallbackImage={null}
-          inCat={inCat}
-          busy={busy}
-          onAdd={onAdd}
-        />
-      ) : null}
-    </View>
-  );
-}
-
-function Checkbox({ checked, disabled }: { checked: boolean; disabled: boolean }) {
-  return (
-    <View style={[styles.cb, checked && styles.cbOn, disabled && styles.cbDisabled]}>
-      {checked && !disabled ? <Ionicons name="checkmark" size={13} color="#fff" /> : null}
-    </View>
-  );
-}
-
-function StatusBadge({ inCat, dup }: { inCat: boolean; dup: boolean }) {
-  if (inCat)
-    return (
-      <View style={styles.badge}>
-        <Ionicons name="checkmark" size={12} color={COLORS.green} />
-        <Text style={[styles.badgeText, { color: COLORS.green }]}>in catalogue</Text>
-      </View>
-    );
-  if (dup)
-    return (
-      <View style={styles.badge}>
-        <Ionicons name="warning" size={12} color={COLORS.yellow} />
-        <Text style={[styles.badgeText, { color: COLORS.yellow }]}>possible dup</Text>
-      </View>
-    );
-  return null;
-}
-
-const styles = StyleSheet.create({
-  modeRow: { flexDirection: 'row', flexWrap: 'wrap', gap: 6, rowGap: 6, marginBottom: 10 },
-  modePill: {
-    paddingHorizontal: 12,
-    paddingVertical: 6,
-    borderRadius: 8,
-    backgroundColor: '#efe6d6',
-  },
-  modePillOn: { backgroundColor: COLORS.navy },
-  modePillText: { fontFamily: 'Nunito_700Bold', fontSize: 12, color: COLORS.navy },
-  modePillTextOn: { color: '#fff' },
-
-  searchBox: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 8,
-    backgroundColor: '#f6f0e6',
-    borderRadius: 10,
-    paddingHorizontal: 13,
-    paddingVertical: 10,
-  },
-  searchInput: { flex: 1, fontFamily: 'Nunito_700Bold', fontSize: 14, color: COLORS.black },
-  popularHint: {
-    fontFamily: 'Nunito_400Regular',
-    fontSize: 12.5,
-    color: COLORS.grey,
-    lineHeight: 18,
-  },
-  loadMore: { alignItems: 'center', justifyContent: 'center', paddingVertical: 12, marginTop: 4 },
-  loadMoreText: {
-    fontFamily: 'Nunito_700Bold',
-    fontSize: 12.5,
-    color: COLORS.orange,
-    textDecorationLine: 'underline',
-  },
-  dim: { opacity: 0.4 },
-
-  toolbar: { flexDirection: 'row', alignItems: 'center', gap: 10, marginTop: 12, flexWrap: 'wrap' },
-  backBtn: { flexDirection: 'row', alignItems: 'center', gap: 3, flex: 1, minWidth: 0 },
-  backText: { fontFamily: 'Nunito_700Bold', fontSize: 13, color: COLORS.navy, flexShrink: 1 },
-  countText: { fontFamily: 'Nunito_700Bold', fontSize: 12, color: COLORS.grey },
-  miniToggle: {
-    paddingHorizontal: 10,
-    paddingVertical: 5,
-    borderRadius: 7,
-    backgroundColor: '#efe6d6',
-  },
-  miniToggleOn: { backgroundColor: COLORS.navy },
-  miniToggleText: { fontFamily: 'Nunito_700Bold', fontSize: 11, color: COLORS.navy },
-  miniToggleTextOn: { color: '#fff' },
-  linkBtn: { paddingVertical: 4 },
-  linkText: {
-    fontFamily: 'Nunito_700Bold',
-    fontSize: 12,
-    color: COLORS.orange,
-    textDecorationLine: 'underline',
-  },
-  addBtn: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 5,
-    backgroundColor: COLORS.orange,
-    borderRadius: 9,
-    paddingHorizontal: 13,
-    paddingVertical: 8,
-  },
-  addBtnText: { fontFamily: 'Nunito_700Bold', fontSize: 12, color: '#fff' },
-  buildAllBtn: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 5,
-    backgroundColor: COLORS.navy,
-    borderRadius: 9,
-    paddingHorizontal: 13,
-    paddingVertical: 8,
-  },
-  buildAllText: { fontFamily: 'Nunito_700Bold', fontSize: 12, color: '#fff' },
-
-  scroll: { maxHeight: 300, marginTop: 8 } as object,
-  charWrap: { borderBottomWidth: 1, borderBottomColor: 'rgba(41,60,67,0.06)' },
-  charWrapOpen: {
-    backgroundColor: '#faf5ec',
-    borderRadius: 10,
-    borderBottomColor: 'transparent',
-    marginVertical: 2,
-  },
-  row: { flexDirection: 'row', alignItems: 'center', gap: 10, paddingVertical: 7 },
-  rowBody: { flex: 1, minWidth: 0, flexDirection: 'row', alignItems: 'center', gap: 10 },
-  groupRow: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 10,
-    paddingVertical: 7,
-    borderBottomWidth: 1,
-    borderBottomColor: 'rgba(41,60,67,0.06)',
-  },
-  groupIcon: {
-    width: 34,
-    height: 44,
-    borderRadius: 7,
-    backgroundColor: COLORS.orange + '14',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  meta: { flex: 1, minWidth: 0, gap: 2 },
-  name: { fontFamily: 'Nunito_700Bold', fontSize: 14, color: COLORS.black },
-  sub: { fontFamily: 'Nunito_400Regular', fontSize: 12, color: COLORS.grey },
-  memberRow: { flexDirection: 'row', alignItems: 'center', gap: 10, paddingVertical: 6 },
-  memberBody: { flex: 1, minWidth: 0, flexDirection: 'row', alignItems: 'center', gap: 10 },
-  memberName: { flex: 1, fontFamily: 'Nunito_400Regular', fontSize: 13, color: COLORS.black },
-  empty: { fontFamily: 'Nunito_400Regular', fontSize: 13, color: COLORS.grey, marginTop: 12 },
-
-  cb: {
-    width: 19,
-    height: 19,
-    borderRadius: 5,
-    borderWidth: 2,
-    borderColor: 'rgba(41,60,67,0.3)',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  cbOn: { backgroundColor: COLORS.orange, borderColor: COLORS.orange },
-  cbDisabled: { borderColor: 'rgba(41,60,67,0.12)', backgroundColor: 'rgba(41,60,67,0.05)' },
-
-  badge: { flexDirection: 'row', alignItems: 'center', gap: 3 },
-  badgeText: { fontFamily: 'Nunito_700Bold', fontSize: 11 },
-
-  session: { marginTop: 12, gap: 8 },
-  loopBar: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 8,
-    backgroundColor: COLORS.orange + '12',
-    borderRadius: 10,
-    paddingHorizontal: 12,
-    paddingVertical: 9,
-  },
-  loopText: { flex: 1, fontFamily: 'Nunito_700Bold', fontSize: 12.5, color: COLORS.navy },
-  loopBtn: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 5,
-    backgroundColor: COLORS.orange,
-    borderRadius: 8,
-    paddingHorizontal: 12,
-    paddingVertical: 7,
-  },
-  loopBtnText: { fontFamily: 'Nunito_700Bold', fontSize: 12, color: '#fff' },
-  addedList: { gap: 2 },
-  addedRow: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 10,
-    paddingVertical: 5,
-    borderBottomWidth: 1,
-    borderBottomColor: 'rgba(41,60,67,0.06)',
-  },
-  addedName: {
-    flex: 1,
-    minWidth: 0,
-    fontFamily: 'Nunito_700Bold',
-    fontSize: 13,
-    color: COLORS.black,
-  },
-  stageBadge: { fontFamily: 'Nunito_700Bold', fontSize: 11, fontVariant: ['tabular-nums'] },
-  removeBtn: {
-    width: 30,
-    height: 30,
-    borderRadius: 8,
-    alignItems: 'center',
-    justifyContent: 'center',
-    backgroundColor: COLORS.red + '12',
-  },
-});
